# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /novels:
    get:
      summary: List all novels
      parameters:
        - in: query
          name: category_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
    post:
      summary: Create new novel (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category_id]
              properties:
                title:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                cover_image:
                  type: string
      responses:
        '201':
          description: Novel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /chapters:
    post:
      summary: Create new chapter (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, title]
              properties:
                novel_id:
                  type: integer
                title:
                  type: string
                order_index:
                  type: integer
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /scenes:
    post:
      summary: Create new scene (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapter_id, title, content]
              properties:
                chapter_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                image_url:
                  type: string
                is_ending:
                  type: boolean
                ending_type:
                  type: string
                  enum: [Good Ending, Bad Ending, True Ending, Secret Ending]
                ending_title:
                  type: string
                ending_description:
                  type: string
      responses:
        '201':
          description: Scene created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    Novel:
      type: object
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
        introduction:
          type: string
        cover_image:
          type: string
        status:
          type: string
        is_published:
          type: boolean
        is_completed:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        views:
          type: integer
        chapter_count:
          type: integer
        scene_count:
          type: integer
        like_count:
          type: integer
        bookshelf_count:
          type: integer
    Chapter:
      type: object
      properties:
        id:
          type: integer
        novel_id:
          type: integer
        title:
          type: string
        order_index:
          type: integer
    Scene:
      type: object
      properties:
        id:
          type: integer
        chapter_id:
          type: integer
        title:
          type: string
        content:
          type: string
        image_url:
          type: string
        is_ending:
          type: boolean
        ending_type:
          type: string
        ending_title:
          type: string
        ending_description:
          type: string
...